+++ src/back/graphInfo.java
@@ -3,7 +3,7 @@ package back;
 import java.util.Vector;
 
 public class graphInfo {
-	public int nr_of_Reads;
+	public double nr_of_Reads;
 	public double avg_slots;
 	public double avg_empty;
 	public double avg_colli;
